Improve group service and event application security
Made modifications to integrate user account for event application and preference creation, ensuring these actions were tied to a specific user. Event application pairing was changed to generate a new pair if none existed or update an existing one. Provided delete capability to preference resolver with added authorization for owner access. Group service updated to retrieve groups based on account id. Added security measures in resolvers assessing user rights. These changes enhance the application's security by preventing unauthorized modifications.

+++ src/app/event/event.resolver.ts
@@ -10,6 +10,7 @@ import {
   Resolver,
 } from '@nestjs/graphql';
 import { Group } from '@prisma/client';
+import { I18nService } from 'nestjs-i18n';
 
 import { Event } from '@/@generated/nestgraphql/event/event.model';
 import { EventApplicationPair } from '@/@generated/nestgraphql/event-application-pair/event-application-pair.model';
@@ -45,29 +46,69 @@ export class EventResolver {
     private readonly eventService: EventService,
     private readonly groupService: GroupService,
     private readonly eventApplicationPairService: EventApplicationPairService,
+    private i18n: I18nService,
   ) {}
+
+  private async checkGroupMembership(
+    accountId: number,
+    groupId: number,
+  ): Promise<void> {
+    const isMember = await this.groupService.ifAccountMemberOfGroup(
+      accountId,
+      groupId,
+    );
+
+    if (!isMember) {
+      throw new Error(this.i18n.t('errors.unauthorized'));
+    }
+  }
 
   @Query(() => [Event], { name: 'events' })
   @UseGuards(AuthGuard)
-  async events(@Args('groupId') groupId: number): Promise<Array<Event> | null> {
+  async events(
+    @Args('groupId') groupId: number,
+    @RequestContextDecorator() context: RequestContext,
+  ): Promise<Array<Event> | null> {
+    await this.checkGroupMembership(context.account!.id, groupId);
     return this.eventService.getEventsByGroupId(groupId);
   }
 
   @Query(() => Event, { name: 'event' })
   @UseGuards(AuthGuard)
-  async event(@Args('id') id: number): Promise<Event | null> {
+  async event(
+    @Args('id') id: number,
+    @RequestContextDecorator() context: RequestContext,
+  ): Promise<Event | null> {
+    const group = await this.groupService.getGroupByEventId(id);
+    if (!group) {
+      // eslint-disable-next-line sonarjs/no-duplicate-string
+      throw new Error(this.i18n.t('errors.notFound'));
+    }
+    await this.checkGroupMembership(context.account!.id, group.id);
     return this.eventService.getEventById(id);
   }
 
   @Mutation(() => Event, { name: 'createEvent' })
   @UseGuards(AuthGuard)
-  async createEvent(@Args('input') input: CreateEventInput): Promise<Event> {
+  async createEvent(
+    @Args('input') input: CreateEventInput,
+    @RequestContextDecorator() context: RequestContext,
+  ): Promise<Event> {
+    await this.checkGroupMembership(context.account!.id, input.groupId);
     return this.eventService.createEvent(input);
   }
 
   @Mutation(() => Boolean, { name: 'deleteEvent' })
   @UseGuards(AuthGuard)
-  async deleteEvent(@Args('id') id: number): Promise<Event> {
+  async deleteEvent(
+    @Args('id') id: number,
+    @RequestContextDecorator() context: RequestContext,
+  ): Promise<Event> {
+    const group = await this.groupService.getGroupByEventId(id);
+    if (!group) {
+      throw new Error(this.i18n.t('errors.notFound'));
+    }
+    await this.checkGroupMembership(context.account!.id, group.id);
     const isDeletable = await this.eventService.isEventDeletable(id);
     if (isDeletable) {
       return await this.eventService.deleteEvent(id);
@@ -82,6 +123,11 @@ export class EventResolver {
     @RequestContextDecorator() context: RequestContext,
     @Parent() event: Event,
   ): Promise<Group | null> {
+    const group = await this.groupService.getGroupByEventId(event.groupId);
+    if (!group) {
+      throw new Error(this.i18n.t('errors.notFound'));
+    }
+    await this.checkGroupMembership(context.account!.id, group.id);
     return this.groupService.getGroupByAccountId(
       context.account!.id,
       event.groupId,
@@ -91,8 +137,14 @@ export class EventResolver {
   @ResolveField(() => [EventApplicationPair], { name: 'eventApplicationPairs' })
   @UseGuards(AuthGuard)
   async eventApplicationPairs(
+    @RequestContextDecorator() context: RequestContext,
     @Parent() event: Event,
   ): Promise<Array<EventApplicationPair> | null> {
+    const group = await this.groupService.getGroupByEventId(event.groupId);
+    if (!group) {
+      throw new Error(this.i18n.t('errors.notFound'));
+    }
+    await this.checkGroupMembership(context.account!.id, group.id);
     return this.eventApplicationPairService.getEventApplicationPairByEventId(
       event.id,
     );
